+++ pyabc/sampler/multicore_evaluation_parallel.py
@@ -5,7 +5,7 @@ from ..sge import nr_cores_available
 import numpy as np
 import random
 from .base import Sample
-
+from .multicorebase import get_if_worker_healthy
 
 DONE = "Done"
 
@@ -101,7 +101,7 @@ class MulticoreEvalParallelSampler(MultiCoreSampler):
         # and the queue is emptied to prevent deadlocks
         n_done = 0
         while n_done < len(processes):
-            val = queue.get()
+            val = get_if_worker_healthy(processes, queue)
             if val == DONE:
                 n_done += 1
             else: